Buildable but with issues
This builds however I need to do some additional verification and testing. 
I am pretty sure there is a silent crash happening and I need to double check to make sure that all functions are properly integrated and functional so this is NOT intended  to be used in a production type environment.