Revisons and security fixes.
Multiple files changed to enhance the security of the code itself. 
Drastic interface changes and some functionality changes. 
Still not quite happy with it where it is but it does build both manually and using the setup_and_run.sh
There is one error on my machine because I have yet to install the application for the GEOIP functions but it does not prevent it from compiling. and otherwise running.
It is fully integrated with ClamAV, Snort / Suricata, and it does actively check for and attempt to handshake to verify select exploits defined within the code.
I am still not 100% satisfied with this implementation but it is a good shell for someone to mess with if they so desire.