Update ReadMe with buildable files status and warnings
Update regarding Buildable but with issues that still remain within the project. The files with different commit messages likely are no longer necessary for the application to function as fully intended but because I have not had a chance to comprehensively test this application I opted to leave them in place.